done upload with txt file

+++ application/controllers/Welcome.php
@@ -22,22 +22,57 @@ class Welcome extends CI_Controller {
 	{
 		if ($this->input->is_ajax_request()) {
 			if ($this->input->get('method') == 'fetch') {
-				$peserta = $this->db->get('peserta')->result();
+				$peserta = $this->db->get('peserta')->result();				
 				echo json_encode($peserta);
 			}else{
 
-				// $listName = explode('\n', $this->input->get('value'));
-				$this->db->query("DELETE FROM peserta where pemenang_keberapa is null");
-				$pemenang = $this->db->query("SELECT * FROM peserta where pemenang_keberapa = 1")->row();
-				$nama_pemenang = $pemenang->first_name;
-				$listName = preg_split('/\r\n|[\r\n]/', $this->input->get('value'));
-				foreach ($listName as $value) {
-					if ($value != "" && $value != $nama_pemenang) {
-						$this->db->query("INSERT INTO peserta(first_name) VALUES ('".$value."')");
-					}					
+				$config['upload_path'] = './uploads/';
+				$config['allowed_types'] = 'txt';
+				// $config['max_size']  = '1000';
+				// $config['max_width']  = '1024';
+				// $config['max_height']  = '768';
+
+				$this->upload->initialize($config);
+				// $this->load->library('upload', $config);
+				if ( ! $this->upload->do_upload('file')){
+					$status = "error";
+					$msg = $this->upload->display_errors();
+					$arr = [
+						'code' => 400,
+						'message' => $msg
+					];
+					echo json_encode($arr);
+					die;
+				}else{
+					$dataupload = $this->upload->data();
+					$status = "success";
+					$msg = $dataupload['file_name']." berhasil diupload";
+
+					// $lines = file('uploads/'.$dataupload['file_name']);
+					$contents = file_get_contents('uploads/'.$dataupload['file_name']);
+					$lines = explode(("\n"), $contents);
+					$this->db->query("DELETE FROM peserta where pemenang_keberapa is null");
+					foreach ($lines as $value)
+					{
+						$this->db->query("INSERT INTO peserta SET first_name=?", [$value]);
+					}
+					$peserta2 = $this->db->get('peserta')->result();
+					echo json_encode($peserta2);
 				}
-				$peserta2 = $this->db->get('peserta')->result();
-				echo json_encode($peserta2);
+				// $this->output->set_content_type('application/json')->set_output(json_encode(array('status'=>$status,'msg'=>$msg)));
+
+
+				// $this->db->query("DELETE FROM peserta where pemenang_keberapa is null");
+				// $pemenang = $this->db->query("SELECT * FROM peserta where pemenang_keberapa = 1")->row();
+				// $nama_pemenang = $pemenang->first_name;
+				// $listName = preg_split('/\r\n|[\r\n]/', $this->input->get('value'));
+				// foreach ($listName as $value) {
+				// 	if ($value != "" && $value != $nama_pemenang) {
+				// 		$this->db->query("INSERT INTO peserta(first_name) VALUES ('".$value."')");
+				// 	}					
+				// }
+				// $peserta2 = $this->db->get('peserta')->result();
+				// echo json_encode($peserta2);
 			}
 		}else{
 			$this->db->select('first_name AS nama,pemenang_keberapa');
